Implemented toml parsing; looks like i will have to remove references from structs in unit.
Implemented toml parsing using serde.
It looks like it is not possible in rust to make references as fields of struct
that reference other parts of that struct, since when you move the
value to the struct the reference becomes invalid. It might be possible
to do with some wierd updates after the struct is created but it is just
simpler to use hash maps and save key values.